+++ CHANGELOG.md
@@ -6,4 +6,5 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
-- Initial commit
+- init: 🎉 initial commit
+- feat: ✨ add toogle and volume commands and properties